+++ tools/src/test/java/org/apache/kafka/tools/LeaderElectionCommandTest.java
@@ -28,7 +28,6 @@ import org.apache.kafka.common.test.ClusterInstance;
 import org.apache.kafka.common.test.api.ClusterConfigProperty;
 import org.apache.kafka.common.test.api.ClusterTest;
 import org.apache.kafka.common.test.api.ClusterTestDefaults;
-import org.apache.kafka.server.common.AdminCommandFailedException;
 
 import org.mockito.ArgumentCaptor;
 import org.mockito.MockedStatic;
